@@ -36,14 +36,14 @@
     "karma-sourcemap-loader": "^0.3.2",
     "karma-webpack": "^1.3.1",
     "mocha": "^2.0.1",
-    "react": "0.12.x",
+    "react": "0.13.0-rc1",
     "rf-changelog": "^0.4.0",
     "rx": "2.3.18",
     "webpack": "^1.4.13",
     "webpack-dev-server": "^1.6.6"
   },
   "peerDependencies": {
-    "react": "0.12.x"
+    "react": "0.13.0-rc1"
   },
   "dependencies": {
     "qs": "2.3.3"
